requirements.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.2
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dlib==19.24.0
docutils==0.19
et-xmlfile==1.1.0
filelock==3.9.0
fire==0.5.0
Flask==2.2.3
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gdown==4.6.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
keras==2.11.0
keyring==23.13.1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplot==0.1.9
matplotlib==3.7.0
mdurl==0.1.2
more-itertools==9.0.0
mtcnn==0.1.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkginfo==1.9.6
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyloco==0.0.139
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-version==0.0.2
pytz==2022.7.1
readme-renderer==37.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retina-face==0.0.13
rfc3986==2.0.0
rich==13.3.1
rsa==4.9
scipy==1.10.0
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
soupsieve==2.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tqdm==4.64.1
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.14
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.13.0
